DACCESS-739 Fix more search errors for BL8
